% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akm.R
\name{akm.connected_set}
\alias{akm.connected_set}
\title{Extract the largest connected set from a data}
\usage{
akm.connected_set(data, ncat = 1, threshold = 0.9)
}
\description{
Extract the largest connected set from a data
}
