% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akm.R
\name{akm.extract_statistics}
\alias{akm.extract_statistics}
\title{Function to extract AKM-related statistics}
\usage{
akm.extract_statistics(wage_data, outcome_var, outcome_predicted = NA)
}
\arguments{
\item{wage_data}{A data.table with information on workers, their wages, which
firm they work at, and the time period. There should be columns called
"firm_ID" and "worker_ID" (data.table).}

\item{outcome_var}{Wage measure on which AKM was computed (e.g., residualized).}

\item{outcome_predicted}{If there is a predicted outcome in the wage_data dataset,
this is the name of that predicted outcome (character).}
}
\description{
This function extracts AKM-related statistics, including variance of
firm effects, variance of worker effects, covariance between firm and
worker effects
}
