% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akm.R
\name{akm.subset_unique_movers}
\alias{akm.subset_unique_movers}
\title{Function to subset firms with a miniumum number of unique movers}
\usage{
akm.subset_unique_movers(wage_data, min_unique_movers)
}
\arguments{
\item{wage_data}{A data.table containing worker IDs and the firm IDs where
the workers are employed. The corresponding columns of the data.table
should be called "worker_ID" and "firm_ID" (data.table).}

\item{min_unique_movers}{The minimum number of unique movers
indicating the number of unique movers a firm must have to remain
in the sample for the estimation (numeric).}
}
\description{
This function takes in data on matched worker and firm ID, counts
the number of unique movers within a firm, then returns the IDs
of firms with a sufficient number of movers and the workers who move.
}
