% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akm.R
\name{akm.weighted_mean}
\alias{akm.weighted_mean}
\title{Compute the weighted mean}
\usage{
akm.weighted_mean(variable_of_interest, weights)
}
\arguments{
\item{variable_of_interest}{Variable to take the mean of (numeric).}

\item{weights}{Weights, which will be rescaled to sum to one (numeric).}
}
\description{
Compute the weighted mean
}
