% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akm.R
\name{akm.zig_zag}
\alias{akm.zig_zag}
\title{Function to estimate the AKM firm effects using the "zig zag" approach.}
\usage{
akm.zig_zag(
  wage_data,
  outcome_var,
  max_iterations = 100,
  tol = 1e-20,
  report_frequency = 1
)
}
\arguments{
\item{wage_data}{A data.table containing worker IDs, the firm IDs where
the workers are employed, the outcome variable of interest, and a time variable.
The corresponding columns of the data.table should be called "worker_ID",
"firm_ID", "year", and the name of the column containing the outcome
variable of interest (see outcome_var argument) (data.table).}

\item{outcome_var}{The name of the column in the wage_data data.table that
contains the outcome variable of interest (character).}

\item{max_iterations}{Maximum number of "zig zag" iterations to use in the
AKM estimation; breaks when max_iterations number of iterations is
completed (numeric).}

\item{tol}{Tolerance for convergence of "zig zag" algorithm AKM estimation;
breaks when tol level of mean squared deviation in the worker effects plus
mean squared deviation in the firm effects is reached (numeric).}

\item{report_frequency}{Determines how many iterations to wait between
printing output. In particular, report_frequency=k means to print a report
every k iterations (numeric).}
}
\description{
This function takes in data on workers, firms, and wages to estimate
the AKM firm effects on the workers' wages. It uses the "zig zag" approach,
in which
}
