% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{check.frame}
\alias{check.frame}
\title{Checking the columns of a data.frame}
\usage{
check.frame(dd, requires = c(), uses = c(), returns = c())
}
\arguments{
\item{requires}{are columns that are required by the function}

\item{uses}{are temporary columns the data.frame will create}

\item{returns}{are the outputs}
}
\description{
Checking the columns of a data.frame
}
