% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeSpellTime}
\alias{computeSpellTime}
\title{Computes spell length within unit}
\usage{
computeSpellTime(unit_id, time, stayer_var)
}
\arguments{
\item{unit_id}{An integer vector with the ID of the unit we consider}

\item{time}{An integer vector with the time considered}

\item{stayer_var}{An integer with the variable that needs to stay constant}
}
\description{
Within unit_id, this will create a counter that updates
when the variable stayer_var changes for when a time is missing
}
