% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE.R
\name{lms.application.blm.subsample}
\alias{lms.application.blm.subsample}
\title{BLM subsample application, subsampling for limited mobility bias analysis}
\usage{
lms.application.blm.subsample(
  full_data,
  min_unique_movers = 15,
  firmclusters = 10,
  flatage = 40,
  residualize = T,
  cluster_sdata_only = F,
  shares = c(0.25, 0.5, 0.75),
  num_random = 2,
  num_cores = 2
)
}
\description{
BLM subsample application, subsampling for limited mobility bias analysis
}
