% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.block_sample}
\alias{lms.block_sample}
\title{draw a block-bootstrap sample (donated by Michael Graber)}
\usage{
lms.block_sample(data, id, time)
}
\arguments{
\item{data}{panel in data.table format}

\item{id}{character; indicating the column name of the individual index}

\item{time}{character; indicating the column name of the time index}
}
\value{
\preformatted{   a data.table with a block bootstrap sample
}
}
\description{
\code{lms.block_sample} draws a block-bootstrap sample from a given panel data set
}
