% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE.R
\name{lms.fe.akm}
\alias{lms.fe.akm}
\title{Main AKM application}
\usage{
lms.fe.akm(
  full_data,
  flatage = 40,
  min_unique_movers = 2,
  tol = 0,
  iter = 100,
  residualize = T,
  va_adjust = F
)
}
\description{
Main AKM application
}
