% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE.R
\name{lms.fe.akm.subsample}
\alias{lms.fe.akm.subsample}
\title{Main AKM application, subsampling for limited mobility bias analysis}
\usage{
lms.fe.akm.subsample(
  full_data,
  flatage = 40,
  min_unique_movers = 15,
  iter = 100,
  tol = 0,
  residualize = T,
  shares = c(0.25, 0.5, 0.75),
  num_random = 2,
  num_cores = 1
)
}
\description{
Main AKM application, subsampling for limited mobility bias analysis
}
