% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE.R
\name{lms.fe.blm}
\alias{lms.fe.blm}
\title{Main BLM}
\usage{
lms.fe.blm(
  full_data,
  min_unique_movers = 2,
  firmclusters = 10,
  flatage = 40,
  residualize = T,
  return_preclustered_data = F,
  cluster_sdata_only = F,
  three_and_three = FALSE,
  va_adjust = FALSE
)
}
\description{
Main BLM
}
