% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE.R
\name{lms.fe.cluster_markets}
\alias{lms.fe.cluster_markets}
\title{Cluster markets and group worker effects by quantile bins}
\usage{
lms.fe.cluster_markets(long_data, marketclusters = 10)
}
\description{
Cluster markets and group worker effects by quantile bins
}
