% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model.extract_ct_moments}
\alias{lms.model.extract_ct_moments}
\title{Model counterfactual moments}
\usage{
lms.model.extract_ct_moments(fxdata, G_constant, price_takers = FALSE)
}
\description{
Model counterfactual moments
}
