% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model.h_fit_grid}
\alias{lms.model.h_fit_grid}
\title{Compare true and predicted values across size distribution}
\usage{
lms.model.h_fit_grid(fxdata, fxdata_solved)
}
\description{
Compare true and predicted values across size distribution
}
