% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model.market_decomp}
\alias{lms.model.market_decomp}
\title{AKM-based decomposition}
\usage{
lms.model.market_decomp(long_data, method = 0)
}
\description{
Construct the different elements for a variance decomposition including market level values. It takes as inputs
inputs: market, year, y_jt, psi_tilde_j, beta_full, alpha_r, rho_r
where psi_tilde_j is the firm fixed effect coming out of an AKM regression on adjusted wages.
outputs: psi_jt, ...
}
\details{
the method argument chooses the approach to construct the time varying part
method 0 uses time variation in y_jt
method 1 uses time variation in l_jt
method 2 uses psi_jt as an input
}
