% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model.prepare}
\alias{lms.model.prepare}
\title{Collect terms required for model decompositions}
\usage{
lms.model.prepare(
  full_data,
  lambda = 0.9154762,
  marketclusters = 10,
  workergroups = 10
)
}
\description{
Collect terms required for model decompositions
}
