% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model}
\alias{lms.model}
\title{All of the model-based results from LMS}
\usage{
lms.model(dd, relax = 0.9, focal_year = 2005)
}
\arguments{
\item{dd}{input data produced by lms.application.blm}

\item{relax}{convergence tolerance parameter}

\item{focal_year}{select a year on which to estimate model counterfactuals}
}
\description{
All of the model-based results from LMS
}
