% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model.solve_Hj}
\alias{lms.model.solve_Hj}
\title{Solving for Hj as a function of the fixed point given
primitives of the model. Requires a few columns from fxdata: GjX,a_rt,a_tilde....}
\usage{
lms.model.solve_Hj(
  fxdata,
  num_steps = 10,
  threshold,
  relax = 0,
  price_takers = FALSE
)
}
\arguments{
\item{relax}{set to 0 for no relaxing, 0.9 is recommended}
}
\description{
Solving for Hj as a function of the fixed point given
primitives of the model. Requires a few columns from fxdata: GjX,a_rt,a_tilde....
}
