% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{lms.model.solve_theta_ct}
\alias{lms.model.solve_theta_ct}
\title{We solve for a counterfactual allocation when we shut shrink the
non-linearity in X in theta_j and G_jX.}
\usage{
lms.model.solve_theta_ct(fxdata, shrink_theta = 0, shrink_g = 0)
}
\arguments{
\item{shrink_theta}{shrinking theta to common value across firms, if 0, nothing happens, if 1, same for everyone}

\item{shrink_g}{shrking the G_jX, if 0 nothing happens, if 1 G_jX is log-additive in j and X}
}
\description{
We solve for a counterfactual allocation when we shut shrink the
non-linearity in X in theta_j and G_jX.
}
