% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.collect_diffInDiffs}
\alias{lms.passthrough.collect_diffInDiffs}
\title{DiD}
\usage{
lms.passthrough.collect_diffInDiffs(wide_data, id_var, firmsize = FALSE)
}
\arguments{
\item{wide_data}{wide_data from lms.passthrough output}

\item{id_var}{identifier}

\item{firmsize}{whether or not it is included}
}
\description{
DiD
}
