% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.collect_IV}
\alias{lms.passthrough.collect_IV}
\title{IV estimation from covariance restrictions}
\usage{
lms.passthrough.collect_IV(
  covmat,
  time_opts = list(spell_length = 8, onset_event_time = 5, forward_event_time = 3,
    backward_event_time = 4)
)
}
\arguments{
\item{covmat}{Covariance matrix produced by lms.passthrough}

\item{time_opts}{Determine the event times to use}
}
\description{
IV estimation from covariance restrictions
}
