% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.collect_params}
\alias{lms.passthrough.collect_params}
\title{Model parameters}
\usage{
lms.passthrough.collect_params(rr, descriptives, rho_to_1 = FALSE)
}
\arguments{
\item{rr}{passthrough results.}

\item{descriptives}{descriptive statistics.}

\item{rho_to_1}{use for the rho=1 counterfactual.}
}
\description{
Model parameters
}
