% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.cov}
\alias{lms.passthrough.cov}
\title{Fast covariances in parallel, returned in long format}
\usage{
lms.passthrough.cov(
  wide_data,
  id_var,
  time_var,
  weight_var = NA,
  number_cores = 1
)
}
\arguments{
\item{wide_data}{Data in wide format.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{weight_var}{Covariance weight variable (optional).}

\item{number_cores}{Number of parallel cores to use for the covariances.}
}
\description{
This takes an input the data structure returned from reshape.longtowide,
which includes leads and lags of variables. It supports parallelization.
It uses the fast covariance function covar from package coop.
}
