% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.covmat_events}
\alias{lms.passthrough.covmat_events}
\title{Infer event times from long covariance matrix}
\usage{
lms.passthrough.covmat_events(
  covmat,
  vars = c("wages", "wagesnet", "wagesmean", "va", "vanet", "vamean", "firmsize",
    "firmsizenet", "firmsizemean")
)
}
\arguments{
\item{covmat}{Covariance matrix produced by lms.passthrough.}

\item{vars}{Variables to include in the event study.}
}
\description{
Infer event times from long covariance matrix
}
