% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.covmat_IV}
\alias{lms.passthrough.covmat_IV}
\title{IV estimation from covariance restrictions}
\usage{
lms.passthrough.covmat_IV(
  covmat,
  spell_length = 8,
  onset_event_time = 5,
  forward_event_time = 3,
  backward_event_time = 4,
  y_name,
  x_name,
  z_name
)
}
\arguments{
\item{covmat}{Covariance matrix produced by lms.passthrough}
}
\description{
IV estimation from covariance restrictions
}
