% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.diffInDiffs_binarize}
\alias{lms.passthrough.diffInDiffs_binarize}
\title{Convert a continuous variable into a binary variable and perform differences-in-differences}
\usage{
lms.passthrough.diffInDiffs_binarize(
  dd,
  id_var,
  time_var,
  timevarying_vars,
  split_var,
  lags,
  number_cores = 1
)
}
\arguments{
\item{dd}{Data in wide format.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{timevarying_vars}{Variables whose response to split_var is to be estimated.}

\item{split_var}{Variable to binarize on its differences.}

\item{lags}{Lags at which the differences have been provided.}

\item{number_cores}{Number of parallel cores to use for the covariances.}
}
\description{
Convert a continuous variable into a binary variable and perform differences-in-differences
}
