% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough}
\alias{lms.passthrough}
\title{Apply all passthrough methods from this package.}
\usage{
lms.passthrough(
  long_data,
  id_var,
  time_var,
  firm_defn,
  market_defn,
  spell_length,
  firm_stayer,
  market_stayer,
  workers_per_firm = NA,
  firms_per_market = NA,
  workers_per_market = NA,
  sample_counts = T,
  sample_info_only = F,
  covmat_numcores = 2,
  outcomes = c("wages", "va", "firmsize"),
  collapse_to_firm = FALSE
)
}
\arguments{
\item{long_data}{Data in long format.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{firm_defn}{Firm identifier.}

\item{market_defn}{Market identifier.}

\item{spell_length}{Length of a spell.}

\item{firm_stayer}{Flag for whether or not workers must stay in same firm over spell_length.}

\item{market_stayer}{Flag for whether or not workers must stay in same market over spell_length.}

\item{workers_per_firm}{Number of id_var spells per firm_defn spells.}

\item{firms_per_market}{Number of firm_defn spells per market_defn spells.}

\item{workers_per_market}{Number of id_var spells per market_defn spells.}

\item{sample_counts}{Collects sample size counts (e.g., unique id_var values) if TRUE.}

\item{sample_info_only}{Returns sample info and skips estimation if TRUE.}

\item{covmat_numcores}{Number of cores to use in computing the covariance matrix.}

\item{outcomes}{Must include "wages". Can also include "firmsize".}

\item{collapse_to_firm}{Indicator for collapsing to firm-level between residualizing and constructing market means.}
}
\description{
Apply all passthrough methods from this package.
}
