% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.reshape_longToWide}
\alias{lms.passthrough.reshape_longToWide}
\title{Reshape data from long to wide}
\usage{
lms.passthrough.reshape_longToWide(
  long_data,
  id_var,
  time_var,
  timevarying_vars,
  constant_vars,
  lags = 1,
  check_adjacent_times = T,
  check_duplicates = T,
  balance_panel = F
)
}
\arguments{
\item{long_data}{Data in long format}

\item{id_var}{ID variable}

\item{time_var}{Time variable}

\item{timevarying_vars}{Variables that we wish to reshape to wide}

\item{constant_vars}{Variables that shouldn't be reshaped because they
do not change with time.}

\item{lags}{Vector of whole numbers indicating the time at which lags
are to be reshaped to wide.}

\item{check_adjacent_times}{Verifies that time_var has increments of 1.}

\item{check_duplicates}{Verify that there are no duplicates in the data.}

\item{balance_panel}{Returns a balanced panel, i.e., use only id_var values
for which an observation is available at all values of lags.}
}
\description{
Reshape data from long to wide
}
