% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.reshape_longToWide_shift}
\alias{lms.passthrough.reshape_longToWide_shift}
\title{Alternative to lms.passthrough.reshape_longToWide with balance=TRUE built on data.table::shift}
\usage{
lms.passthrough.reshape_longToWide_shift(
  long_data,
  id_var,
  time_var,
  timevarying_vars,
  lags
)
}
\arguments{
\item{long_data}{Data in long format.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{timevarying_vars}{Variables that we wish to reshape to wide.}

\item{lags}{Vector of whole numbers indicating the time at which lags
are to be reshaped to wide.}
}
\description{
Alternative to lms.passthrough.reshape_longToWide with balance=TRUE built on data.table::shift
}
