% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.reshape_wideToLong}
\alias{lms.passthrough.reshape_wideToLong}
\title{Reshape data from wide to long}
\usage{
lms.passthrough.reshape_wideToLong(
  wide_data,
  id_var,
  time_var,
  timevarying_vars,
  constant_vars,
  lags = 1
)
}
\arguments{
\item{wide_data}{Data in wide format}

\item{id_var}{ID variable}

\item{time_var}{Time variable}

\item{timevarying_vars}{Variables that we wish to reshape to lon}

\item{constant_vars}{Variables that shouldn't be reshaped because they
do not change with time.}

\item{lags}{Vector of whole numbers indicating the time at which lags
are to be reshaped to wide.}
}
\description{
Reshape data from wide to long
}
