% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.reshape_wideToLong_melt}
\alias{lms.passthrough.reshape_wideToLong_melt}
\title{Alternative to lms.passthrough.reshape_longToWide built on data.table::melt}
\usage{
lms.passthrough.reshape_wideToLong_melt(
  wide_data,
  id_var,
  time_var,
  timevarying_vars
)
}
\arguments{
\item{wide_data}{Data in wide format.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{timevarying_vars}{Variables that we wish to reshape to long.}
}
\description{
Alternative to lms.passthrough.reshape_longToWide built on data.table::melt
}
