% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.select_countPerGroup_wide}
\alias{lms.passthrough.select_countPerGroup_wide}
\title{Select the sample with a minimum count of observations per group.}
\usage{
lms.passthrough.select_countPerGroup_wide(
  wide_data,
  inner_group_var,
  outer_group_vars,
  min_count
)
}
\arguments{
\item{wide_data}{Data in wide format.}

\item{inner_group_var}{Name of variable that defines the inner group.}

\item{outer_group_vars}{Name(s) of variables that define the outer group.}

\item{min_count}{Minimum count per group.}
}
\description{
Select the sample with a minimum count of observations per group.
}
