% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.select_init}
\alias{lms.passthrough.select_init}
\title{Initialize spell data}
\usage{
lms.passthrough.select_init(long_data)
}
\arguments{
\item{long_data}{Data in long format.}
}
\description{
Initialize spell data
}
