% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.select_nonmissing_wide}
\alias{lms.passthrough.select_nonmissing_wide}
\title{Select the sample that is not missing on a given variable and its lags}
\usage{
lms.passthrough.select_nonmissing_wide(
  wide_data,
  nonmissing_var,
  lags,
  current_time = T
)
}
\arguments{
\item{wide_data}{Data in wide format}

\item{nonmissing_var}{Variable with respect to which the observation
must not be missing.}

\item{lags}{Lags at which the observation must be a stayer.}

\item{current_time}{Impose stayer at the current time (i.e., lag=0).}
}
\description{
Select the sample that is not missing on a given variable and its lags
}
