% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.select_stayers_wide}
\alias{lms.passthrough.select_stayers_wide}
\title{Select the sample of stayers on a given variable when data is wide}
\usage{
lms.passthrough.select_stayers_wide(
  wide_data,
  stayer_var,
  lags,
  current_time = T
)
}
\arguments{
\item{wide_data}{Data in wide format}

\item{stayer_var}{Variable with respect to which the observation
must be a stayer (i.e., same observation value across lags).}

\item{lags}{Lags at which the observation must be a stayer.}

\item{current_time}{Impose stayer at the current time (i.e., lag=0).}
}
\description{
Select the sample of stayers on a given variable when data is wide
}
