% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.select_unitCount}
\alias{lms.passthrough.select_unitCount}
\title{Select the sample of stayers on a given variable when data is wide}
\usage{
lms.passthrough.select_unitCount(
  long_data,
  id_var,
  time_var,
  stayer_var,
  spell_length,
  min_units = 1,
  drop = T
)
}
\arguments{
\item{long_data}{Data in long format.}

\item{id_var}{ID variable.}

\item{time_var}{Time variable.}

\item{stayer_var}{Variable with respect to which the observation
must be a stayer (i.e., same value across lags).}

\item{spell_length}{Length of a spell.}

\item{min_units}{Minimum number of id_var spells per stayer_var spell.}

\item{drop}{Drop units that do not satisfy the restriction.}
}
\description{
Select the sample of stayers on a given variable when data is wide
}
