% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.transform_detrendMarkets}
\alias{lms.passthrough.transform_detrendMarkets}
\title{felm-based function to residualize out market-specific linear time trends from a specified variable.}
\usage{
lms.passthrough.transform_detrendMarkets(
  input_data,
  residualize_var,
  time_var,
  market_defn
)
}
\arguments{
\item{input_data}{The data containing the outcome and covariates (data.table).}

\item{residualize_var}{The name of the outcome variable (character).}

\item{time_var}{The name of the time variable (character).}

\item{market_defn}{The definition of a market (character).}
}
\description{
This function regresses an outcome on each market's linear time trend.
}
