% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passthrough.R
\name{lms.passthrough.transform_difference_wide}
\alias{lms.passthrough.transform_difference_wide}
\title{Take first differences at the given lags}
\usage{
lms.passthrough.transform_difference_wide(
  wide_data,
  difference_var,
  lags = 1,
  current_time = T
)
}
\arguments{
\item{wide_data}{Data in long format.}

\item{difference_var}{Variable to difference.}

\item{lags}{Vector of whole numbers indicating the time at which lags
are to be reshaped to wide.}

\item{current_time}{Include the current time period.}
}
\description{
Take first differences at the given lags
}
